Move DatabaseDump into test namespace since it is only used there
