+++ .github/workflows/phpstan/return-types-for-new-class-methods-rule-exclusion-list.php
@@ -733,18 +733,18 @@ $baseline = [
     'PrestaShopBundle\Form\Admin\Improve\International\Geolocation\GeolocationFormDataProvider::__construct',
     'PrestaShopBundle\Form\Admin\Improve\International\Geolocation\GeolocationFormDataProvider::getData',
     'PrestaShopBundle\Form\Admin\Improve\International\Geolocation\GeolocationFormDataProvider::setData',
-    'PrestaShopBundle\Install\DatabaseDump::__construct',
+    'Tests\TestCase\DatabaseDump::__construct',
     'PrestaShopBundle\Form\Admin\Improve\International\Geolocation\GeolocationByIpAddressType::buildForm',
     'PrestaShopBundle\Form\Admin\Improve\Design\Theme\ShopLogosType::disableAllShopContextFields',
-    'PrestaShopBundle\Install\DatabaseDump::buildMySQLCommand',
+    'Tests\TestCase\DatabaseDump::buildMySQLCommand',
     'PrestaShopBundle\Form\Admin\Improve\International\Geolocation\GeolocationOptionsFormDataProvider::__construct',
-    'PrestaShopBundle\Install\DatabaseDump::exec',
+    'Tests\TestCase\DatabaseDump::exec',
     'PrestaShopBundle\Form\Admin\Improve\International\Geolocation\GeolocationOptionsFormDataProvider::getData',
-    'PrestaShopBundle\Install\DatabaseDump::dump',
+    'Tests\TestCase\DatabaseDump::dump',
     'PrestaShopBundle\Form\Admin\Improve\International\Geolocation\GeolocationOptionsFormDataProvider::setData',
-    'PrestaShopBundle\Install\DatabaseDump::restore',
-    'PrestaShopBundle\Install\DatabaseDump::create',
-    'PrestaShopBundle\Install\DatabaseDump::restoreDb',
+    'Tests\TestCase\DatabaseDump::restore',
+    'Tests\TestCase\DatabaseDump::create',
+    'Tests\TestCase\DatabaseDump::restoreDb',
     'PrestaShopBundle\Form\Admin\Improve\International\Geolocation\GeolocationWhitelistFormDataProvider::__construct',
     'PrestaShopBundle\Form\Admin\Improve\International\Geolocation\GeolocationWhitelistFormDataProvider::getData',
     'PrestaShopBundle\Form\Admin\Improve\International\Geolocation\GeolocationWhitelistFormDataProvider::setData',
