+++ .github/workflows/phpstan/type-hint-for-new-class-methods-rule-exclusion-list.php
@@ -95,10 +95,10 @@ $baseline = [
 'PrestaShopBundle\Form\DataTransformer\ArabicToLatinDigitDataTransformer::transform',
 'PrestaShopBundle\Form\DataTransformer\ArabicToLatinDigitDataTransformer::reverseTransform',
 'PrestaShopBundle\Form\Admin\Improve\Design\Theme\ShopLogosType::transformMultiStoreFields',
-'PrestaShopBundle\Install\DatabaseDump::__construct',
-'PrestaShopBundle\Install\DatabaseDump::buildMySQLCommand',
+'Tests\TestCase\DatabaseDump::__construct',
+'Tests\TestCase\DatabaseDump::buildMySQLCommand',
 'PrestaShopBundle\Form\Admin\Improve\Design\Theme\ShopLogosType::disableAllShopContextFields',
-'PrestaShopBundle\Install\DatabaseDump::exec',
+'Tests\TestCase\DatabaseDump::exec',
 'PrestaShopBundle\Form\Admin\Improve\Design\Theme\ShopLogosType::setShopRestrictionSource',
 'PrestaShopBundle\Form\Admin\Improve\Design\Theme\ShopLogosType::getShopRestrictionSourceFormFields',
 'PrestaShopBundle\Form\Admin\Improve\Design\Theme\ShopLogosType::stringEndsWith',
