+++ tests/Integration/Behaviour/Features/Context/CommonFeatureContext.php
@@ -79,7 +79,6 @@ use OrderState;
 use Pack;
 use Page;
 use PrestaShop\PrestaShop\Adapter\LegacyContext;
-use PrestaShopBundle\Install\DatabaseDump;
 use Product;
 use ProductAttribute;
 use ProductDownload;
@@ -116,6 +115,7 @@ use TaxManagerFactory;
 use TaxRule;
 use TaxRulesGroup;
 use Tests\Integration\Utility\ContextMocker;
+use Tests\Resources\DatabaseDump;
 use Tests\Resources\ResourceResetter;
 use WarehouseProductLocation;
 use WebserviceKey;
