+++ tests/Resources/DatabaseDump.php
@@ -25,7 +25,7 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
 
-namespace PrestaShopBundle\Install;
+namespace Tests\Resources;
 
 use AppKernel;
 use Db;
