+++ tests/bin/restore-test-db.php
@@ -25,7 +25,7 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
 
-use PrestaShopBundle\Install\DatabaseDump;
+use Tests\Resources\DatabaseDump;
 
 define('_PS_ROOT_DIR_', dirname(__DIR__, 2));
 const _PS_IN_TEST_ = true;
